Rework models to use zod
