--- webpack.config.babel.mjs
@@ -1,30 +1,0 @@
-import { resolve } from 'path';
-const target = process.env.TARGET || 'all';
-
-const flatten = input => {
-  const stack = [...input];
-  const res = [];
-  while (stack.length) {
-    const next = stack.pop();
-    if (Array.isArray(next)) {
-      stack.push(...next);
-    } else {
-      res.push(next);
-    }
-  }
-  return res.reverse();
-};
-
-const aquireTarget = inTarget => require(resolve(__dirname, 'webpack', inTarget, `index.mjs`))
-
-const configuration =
-  target === 'all'
-    ? [
-        aquireTarget('style').config,
-        aquireTarget('client').config,
-        aquireTarget('server').config,
-      ]
-    : [aquireTarget(target).config];
-
-
-export default [...flatten(configuration)];
