+++ webpack/common.js
